# RegexGroupParser - cache regex parsing results