running https://github.com/phpstan/phpstan-src/blob/327ac3e54bd04d85aec3af9f893f8e3e8f38af7d/tests/PHPStan/Analyser/nsrt/preg_match_shapes.php 

before this PR
```
➜  phpstan-src git:(2.1.x) ✗ hyperfine 'php bin/phpstan analyze test.php --debug' -i
Benchmark 1: php bin/phpstan analyze test.php --debug
  Time (mean ± σ):      1.677 s ±  0.007 s    [User: 1.563 s, System: 0.110 s]
  Range (min … max):    1.666 s …  1.685 s    10 runs
```

after this PR
```
➜  phpstan-src git:(2.1.x) ✗ hyperfine 'php bin/phpstan analyze test.php --debug' -i
Benchmark 1: php bin/phpstan analyze test.php --debug
  Time (mean ± σ):      1.280 s ±  0.018 s    [User: 1.165 s, System: 0.111 s]
  Range (min … max):    1.264 s …  1.311 s    10 runs
```